Update GraphicsTextureResource  #126
Create an enum textureName_e which contains all enumeration of textures.
Update this resource to store a std::map<textureName_e, std::shared_ptrsf::Texture>.
Implement a function which take an enum and a path to add a texture to the map.
The constructor can take as parameter an enum and a texture path.
The destructor can be set as default.